Create ServingConfig and update name
