--- docs/search-core.cloudregion.md
@@ -1,21 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@yext/search-core](./search-core.md) &gt; [CloudRegion](./search-core.cloudregion.md)
-
-## CloudRegion enum
-
-Defines the cloud region of the API domains.
-
-<b>Signature:</b>
-
-```typescript
-export declare enum CloudRegion 
-```
-
-## Enumeration Members
-
-|  Member | Value | Description |
-|  --- | --- | --- |
-|  EU | <code>&quot;eu&quot;</code> |  |
-|  US | <code>&quot;us&quot;</code> |  |
-
