--- docs/search-core.environment.md
@@ -1,21 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@yext/search-core](./search-core.md) &gt; [Environment](./search-core.environment.md)
-
-## Environment enum
-
-Defines the environment of the API domains.
-
-<b>Signature:</b>
-
-```typescript
-export declare enum Environment 
-```
-
-## Enumeration Members
-
-|  Member | Value | Description |
-|  --- | --- | --- |
-|  PROD | <code>&quot;prod&quot;</code> |  |
-|  SANDBOX | <code>&quot;sbx&quot;</code> |  |
-
