+++ docs/search-core.md
@@ -17,10 +17,8 @@
 |  --- | --- |
 |  [AppliedQueryFilterType](./search-core.appliedqueryfiltertype.md) | Represents the type of [AppliedQueryFilter](./search-core.appliedqueryfilter.md) applied to a search. |
 |  [BuiltInFieldType](./search-core.builtinfieldtype.md) | Possible built-in field types for [DirectAnswer.fieldType](./search-core.directanswer.fieldtype.md)<!-- -->. |
-|  [CloudRegion](./search-core.cloudregion.md) | Defines the cloud region of the API domains. |
 |  [DirectAnswerType](./search-core.directanswertype.md) | Represents the type of direct answer. |
 |  [Direction](./search-core.direction.md) | The direction of a sort. |
-|  [Environment](./search-core.environment.md) | Defines the environment of the API domains. |
 |  [ErrorType](./search-core.errortype.md) | Identifier for the type of error causing the failure. |
 |  [FilterCombinator](./search-core.filtercombinator.md) | Indicates how child filters in a [StaticFilter](./search-core.staticfilter.md) should be combined. |
 |  [LocationBiasMethod](./search-core.locationbiasmethod.md) | The method used to determine the location. |
@@ -37,7 +35,7 @@
 |  Function | Description |
 |  --- | --- |
 |  [provideCore(config)](./search-core.providecore.md) | The entrypoint to the search-core library. |
-|  [provideEndpoints(environment, cloudRegion)](./search-core.provideendpoints.md) | Provides all endpoints based on environment and cloud region. |
+|  [provideEndpoints(config)](./search-core.provideendpoints.md) | Provides all endpoints based on environment and cloud region. |
 
 ## Interfaces
 
@@ -103,6 +101,7 @@
 |  [SearchParameterField](./search-core.searchparameterfield.md) | Indicates which entity field to perform the autocomplete request on. |
 |  [SearchRequest](./search-core.searchrequest.md) | Options for a Search API request. |
 |  [SearchService](./search-core.searchservice.md) | A service which performs Yext Search. |
+|  [ServingConfig](./search-core.servingconfig.md) | The configuration options for [provideEndpoints()](./search-core.provideendpoints.md)<!-- -->. |
 |  [Snippet](./search-core.snippet.md) | The section of text where a [FeaturedSnippetDirectAnswer](./search-core.featuredsnippetdirectanswer.md) was found. |
 |  [SortBy](./search-core.sortby.md) | Represents a criterion that can be used to sort results. |
 |  [SpellCheck](./search-core.spellcheck.md) | A spellcheck response from a search query. |
