+++ docs/search-core.servingconfig.cloudregion.md
@@ -1,0 +1,12 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@yext/search-core](./search-core.md) &gt; [ServingConfig](./search-core.servingconfig.md) &gt; [cloudRegion](./search-core.servingconfig.cloudregion.md)
+
+## ServingConfig.cloudRegion property
+
+
+<b>Signature:</b>
+
+```typescript
+cloudRegion?: CloudRegion;
+```
