+++ docs/search-core.servingconfig.environment.md
@@ -1,0 +1,12 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@yext/search-core](./search-core.md) &gt; [ServingConfig](./search-core.servingconfig.md) &gt; [environment](./search-core.servingconfig.environment.md)
+
+## ServingConfig.environment property
+
+
+<b>Signature:</b>
+
+```typescript
+environment?: Environment;
+```
