+++ docs/search-core.servingconfig.md
@@ -1,0 +1,21 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@yext/search-core](./search-core.md) &gt; [ServingConfig](./search-core.servingconfig.md)
+
+## ServingConfig interface
+
+The configuration options for [provideEndpoints()](./search-core.provideendpoints.md)<!-- -->.
+
+<b>Signature:</b>
+
+```typescript
+export interface ServingConfig 
+```
+
+## Properties
+
+|  Property | Type | Description |
+|  --- | --- | --- |
+|  [cloudRegion?](./search-core.servingconfig.cloudregion.md) | CloudRegion | <i>(Optional)</i> |
+|  [environment?](./search-core.servingconfig.environment.md) | Environment | <i>(Optional)</i> |
+
