+++ etc/search-core.api.md
@@ -105,7 +105,7 @@ export interface BaseFieldValueDirectAnswer<T = unknown> extends DirectAnswer<T>
 }
 
 // @public
-export interface BaseSearchConfig {
+export interface BaseSearchConfig extends ServingConfig {
     additionalQueryParams?: {
         [key: string]: string | number | boolean;
     };
@@ -156,14 +156,6 @@ export enum BuiltInFieldType {
 export interface ClientSDKHeaderValues {
     [agent: string]: string | undefined;
     ANSWERS_CORE?: never;
-}
-
-// @public
-export enum CloudRegion {
-    // (undocumented)
-    EU = "eu",
-    // (undocumented)
-    US = "us"
 }
 
 // @public
@@ -273,14 +265,6 @@ export interface Endpoints {
 
 // @public
 export type EnumOrLiteral<T extends string> = T | `${T}`;
-
-// @public
-export enum Environment {
-    // (undocumented)
-    PROD = "prod",
-    // (undocumented)
-    SANDBOX = "sbx"
-}
 
 // @public
 export enum ErrorType {
@@ -528,7 +512,7 @@ export interface PhoneDirectAnswer extends BaseFieldValueDirectAnswer<string> {
 export function provideCore(config: SearchConfig): SearchCore;
 
 // @public
-export function provideEndpoints(environment?: Environment, cloudRegion?: CloudRegion): Required<Endpoints>;
+export function provideEndpoints(config?: ServingConfig): Required<Endpoints>;
 
 // @public
 export interface QueryRulesActionsData {
@@ -659,6 +643,20 @@ export interface SearchRequest {
 export interface SearchService {
     universalSearch(request: UniversalSearchRequest): Promise<UniversalSearchResponse>;
     verticalSearch(request: VerticalSearchRequest): Promise<VerticalSearchResponse>;
+}
+
+// @public
+export interface ServingConfig {
+    // Warning: (ae-forgotten-export) The symbol "CloudRegion" needs to be exported by the entry point index.d.ts
+    // Warning: (ae-unresolved-inheritdoc-reference) The @inheritDoc reference could not be resolved: The package "@yext/search-core" does not have an export "CloudRegion"
+    //
+    // (undocumented)
+    cloudRegion?: CloudRegion;
+    // Warning: (ae-forgotten-export) The symbol "Environment" needs to be exported by the entry point index.d.ts
+    // Warning: (ae-unresolved-inheritdoc-reference) The @inheritDoc reference could not be resolved: The package "@yext/search-core" does not have an export "Environment"
+    //
+    // (undocumented)
+    environment?: Environment;
 }
 
 // @public
