+++ src/models/core/CloudRegion.ts
@@ -1,0 +1,9 @@
+/**
+ * Defines the cloud region of the API domains.
+ *
+ * @public
+ */
+export enum CloudRegion {
+  US = 'us',
+  EU = 'eu',
+}
