+++ src/models/core/Environment.ts
@@ -1,0 +1,9 @@
+/**
+ * Defines the environment of the API domains.
+ *
+ * @public
+ */
+export enum Environment {
+  PROD = 'prod',
+  SANDBOX = 'sbx',
+}
